Refactor download to module
- Add more UT coverage
