+++ tests/test_blobxfer_models.py
@@ -348,159 +348,3 @@ def test_azurestorageentity():
     ase.populate_from_file(mock.MagicMock(), blob)
     assert ase.mode == models.AzureStorageModes.File
     assert ase.snapshot is None
-
-
-def test_downloaddescriptor(tmpdir):
-    lp = pathlib.Path(str(tmpdir.join('a')))
-
-    opts = mock.MagicMock()
-    opts.check_file_md5 = True
-    opts.chunk_size_bytes = 16
-    ase = models.AzureStorageEntity('cont')
-    ase._size = 1024
-    ase._encryption = mock.MagicMock()
-    with pytest.raises(RuntimeError):
-        d = models.DownloadDescriptor(lp, ase, opts)
-
-    ase._encryption.symmetric_key = b'123'
-    d = models.DownloadDescriptor(lp, ase, opts)
-
-    assert d.entity == ase
-    assert not d.must_compute_md5
-    assert d._total_chunks == 64
-    assert d._offset == 0
-    assert d.final_path == lp
-    assert str(d.local_path) == str(lp) + '.bxtmp'
-    assert d.local_path.stat().st_size == 1024 - 16
-
-    d.local_path.unlink()
-    ase._size = 1
-    d = models.DownloadDescriptor(lp, ase, opts)
-    assert d._total_chunks == 1
-    assert d.local_path.stat().st_size == 0
-
-    d.local_path.unlink()
-    ase._encryption = None
-    ase._size = 1024
-    d = models.DownloadDescriptor(lp, ase, opts)
-    assert d.local_path.stat().st_size == 1024
-
-    # pre-existing file check
-    ase._size = 0
-    d = models.DownloadDescriptor(lp, ase, opts)
-    assert d._total_chunks == 0
-    assert d.local_path.stat().st_size == 0
-
-
-def test_downloaddescriptor_next_offsets(tmpdir):
-    lp = pathlib.Path(str(tmpdir.join('a')))
-
-    opts = mock.MagicMock()
-    opts.check_file_md5 = True
-    opts.chunk_size_bytes = 256
-    ase = models.AzureStorageEntity('cont')
-    ase._size = 128
-    d = models.DownloadDescriptor(lp, ase, opts)
-
-    offsets = d.next_offsets()
-    assert d._total_chunks == 1
-    assert offsets.chunk_num == 0
-    assert offsets.fd_start == 0
-    assert offsets.num_bytes == 128
-    assert offsets.range_start == 0
-    assert offsets.range_end == 127
-    assert not offsets.unpad
-    assert d.next_offsets() is None
-
-    ase._size = 0
-    d = models.DownloadDescriptor(lp, ase, opts)
-    assert d._total_chunks == 0
-    assert d.next_offsets() is None
-
-    ase._size = 1
-    d = models.DownloadDescriptor(lp, ase, opts)
-    offsets = d.next_offsets()
-    assert d._total_chunks == 1
-    assert offsets.chunk_num == 0
-    assert offsets.fd_start == 0
-    assert offsets.num_bytes == 1
-    assert offsets.range_start == 0
-    assert offsets.range_end == 0
-    assert not offsets.unpad
-    assert d.next_offsets() is None
-
-    ase._size = 256
-    d = models.DownloadDescriptor(lp, ase, opts)
-    offsets = d.next_offsets()
-    assert d._total_chunks == 1
-    assert offsets.chunk_num == 0
-    assert offsets.fd_start == 0
-    assert offsets.num_bytes == 256
-    assert offsets.range_start == 0
-    assert offsets.range_end == 255
-    assert not offsets.unpad
-    assert d.next_offsets() is None
-
-    ase._size = 256 + 16
-    d = models.DownloadDescriptor(lp, ase, opts)
-    offsets = d.next_offsets()
-    assert d._total_chunks == 2
-    assert offsets.chunk_num == 0
-    assert offsets.fd_start == 0
-    assert offsets.num_bytes == 256
-    assert offsets.range_start == 0
-    assert offsets.range_end == 255
-    assert not offsets.unpad
-    offsets = d.next_offsets()
-    assert offsets.chunk_num == 1
-    assert offsets.fd_start == 256
-    assert offsets.num_bytes == 16
-    assert offsets.range_start == 256
-    assert offsets.range_end == 256 + 15
-    assert not offsets.unpad
-    assert d.next_offsets() is None
-
-    ase._encryption = mock.MagicMock()
-    ase._encryption.symmetric_key = b'123'
-    ase._size = 128
-    d = models.DownloadDescriptor(lp, ase, opts)
-    offsets = d.next_offsets()
-    assert d._total_chunks == 1
-    assert offsets.chunk_num == 0
-    assert offsets.fd_start == 0
-    assert offsets.num_bytes == 128
-    assert offsets.range_start == 0
-    assert offsets.range_end == 127
-    assert offsets.unpad
-    assert d.next_offsets() is None
-
-    ase._size = 256
-    d = models.DownloadDescriptor(lp, ase, opts)
-    offsets = d.next_offsets()
-    assert d._total_chunks == 1
-    assert offsets.chunk_num == 0
-    assert offsets.fd_start == 0
-    assert offsets.num_bytes == 256
-    assert offsets.range_start == 0
-    assert offsets.range_end == 255
-    assert offsets.unpad
-    assert d.next_offsets() is None
-
-    ase._size = 256 + 32  # 16 bytes over + padding
-    d = models.DownloadDescriptor(lp, ase, opts)
-    offsets = d.next_offsets()
-    assert d._total_chunks == 2
-    assert offsets.chunk_num == 0
-    assert offsets.fd_start == 0
-    assert offsets.num_bytes == 256
-    assert offsets.range_start == 0
-    assert offsets.range_end == 255
-    assert not offsets.unpad
-    offsets = d.next_offsets()
-    assert offsets.chunk_num == 1
-    assert offsets.fd_start == 256
-    assert offsets.num_bytes == 32
-    assert offsets.range_start == 256 - 16
-    assert offsets.range_end == 256 + 31
-    assert offsets.unpad
-    assert d.next_offsets() is None
